+++ crates/ruma-events/src/room.rs
@@ -41,6 +41,8 @@ pub mod name;
 pub mod pinned_events;
 pub mod policy;
 pub mod power_levels;
+#[cfg(feature = "unstable-msc4495")]
+pub mod presence_sharing;
 pub mod redaction;
 pub mod server_acl;
 pub mod third_party_invite;